Move lsp/ to plugins (with no regard for breaking things)
