--- python/src/aac/lang/lsp/__init__.py
@@ -1,1 +1,0 @@
-"""Package for Language Server Protocol specific modules."""
